Conflicts:
	nds-client/src/main/java/br/com/abril/nds/repository/CotaRepository.java
	nds-client/src/main/java/br/com/abril/nds/repository/impl/CotaRepositoryImpl.java
	nds-client/src/main/java/br/com/abril/nds/service/impl/CotaServiceImpl.java

+++ nds-client/src/main/java/br/com/abril/nds/service/impl/FiadorServiceImpl.java
@@ -2,6 +2,7 @@ package br.com.abril.nds.service.impl;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
@@ -165,6 +166,57 @@ public class FiadorServiceImpl implements FiadorService {
 			}
 		}
 		
+		if (fiador.getPessoa() instanceof PessoaFisica){
+			
+			Long idPessoa = this.pessoaRepository.buscarIdPessoaPorCPF(((PessoaFisica) fiador.getPessoa()).getCpf());
+			fiador.getPessoa().setId(idPessoa);
+		} else {
+			
+			Long idPessoa = this.pessoaRepository.buscarIdPessoaPorCNPJ(((PessoaJuridica) fiador.getPessoa()).getCnpj());
+			fiador.getPessoa().setId(idPessoa);
+		}
+		
+		if (fiador.getPessoa() instanceof PessoaFisica){
+			if (((PessoaFisica) fiador.getPessoa()).getConjuge() != null){
+				((PessoaFisica) fiador.getPessoa()).getConjuge().setConjuge((PessoaFisica) fiador.getPessoa());
+			}
+		}
+		
+		if (fiador.getPessoa() instanceof PessoaJuridica){
+			this.processarSocios(fiador, sociosAdicionar, sociosRemover);
+		}
+		
+		if (fiador.getPessoa().getId() == null){
+			
+			this.pessoaRepository.adicionar(fiador.getPessoa());
+		} else {
+			
+			this.pessoaRepository.alterar(fiador.getPessoa());
+		}
+		
+		if (fiador.getId() == null){
+			
+			fiador.setInicioAtividade(new Date());
+			
+			this.fiadorRepository.adicionar(fiador);
+		} else {
+			
+			fiador.setInicioAtividade(this.fiadorRepository.buscarDataInicioAtividadeFiadorPorId(fiador.getId()));
+			
+			this.fiadorRepository.alterar(fiador);
+		}
+		
+		this.processarEnderecos(fiador, listaEnderecosAdicionar, listaEnderecosRemover);
+		
+		this.processarTelefones(fiador, listaTelefoneAdicionar, listaTelefoneRemover);
+		
+		this.processarGarantias(fiador, listaGarantiaAdicionar, listaGarantiaRemover);
+		
+		this.processarCotasAssociadas(fiador, listaCotas, listaCotasDesassociar);
+	}
+
+	private void processarSocios(Fiador fiador, List<Pessoa> sociosAdicionar, Set<Long> sociosRemover) {
+		
 		if (sociosAdicionar != null && !sociosAdicionar.isEmpty()){
 			
 			for (Pessoa socio : sociosAdicionar){
@@ -175,18 +227,27 @@ public class FiadorServiceImpl implements FiadorService {
 						
 						PessoaFisica conjuge = ((PessoaFisica) socio).getConjuge();
 						
+						Long idPessoa = this.pessoaRepository.buscarIdPessoaPorCPF(conjuge.getCpf());
+						conjuge.setId(idPessoa);
+						
 						if (conjuge.getId() == null){
+							
 							this.pessoaRepository.adicionar(conjuge);
 						} else {
+							
 							this.pessoaRepository.alterar(conjuge);
 						}
 					}
+					
+					Long idPessoa = this.pessoaRepository.buscarIdPessoaPorCPF(((PessoaFisica) socio).getCpf());
+					socio.setId(idPessoa);
 				}
 				
-				
 				if (socio.getId() == null){
+					
 					this.pessoaRepository.adicionar(socio);
 				} else {
+					
 					this.pessoaRepository.alterar(socio);
 				}
 			}
@@ -223,46 +284,6 @@ public class FiadorServiceImpl implements FiadorService {
 			
 			fiador.setSocios(sociosAdicionar);
 		}
-		
-		if (fiador.getPessoa() instanceof PessoaFisica){
-			
-			Long idPessoa = this.pessoaRepository.buscarIdPessoaPorCPF(((PessoaFisica) fiador.getPessoa()).getCpf());
-			fiador.getPessoa().setId(idPessoa);
-		} else {
-			
-			Long idPessoa = this.pessoaRepository.buscarIdPessoaPorCPF(((PessoaJuridica) fiador.getPessoa()).getCnpj());
-			fiador.getPessoa().setId(idPessoa);
-		}
-		
-		if (fiador.getPessoa() instanceof PessoaFisica){
-			if (((PessoaFisica) fiador.getPessoa()).getConjuge() != null){
-				((PessoaFisica) fiador.getPessoa()).getConjuge().setConjuge((PessoaFisica) fiador.getPessoa());
-			}
-		}
-		
-		if (fiador.getPessoa().getId() == null){
-			
-			this.pessoaRepository.adicionar(fiador.getPessoa());
-		} else {
-			
-			this.pessoaRepository.alterar(fiador.getPessoa());
-		}
-		
-		if (fiador.getId() == null){
-			
-			this.fiadorRepository.adicionar(fiador);
-		} else {
-			
-			this.fiadorRepository.alterar(fiador);
-		}
-		
-		this.processarEnderecos(fiador, listaEnderecosAdicionar, listaEnderecosRemover);
-		
-		this.processarTelefones(fiador, listaTelefoneAdicionar, listaTelefoneRemover);
-		
-		this.processarGarantias(fiador, listaGarantiaAdicionar, listaGarantiaRemover);
-		
-		this.processarCotasAssociadas(fiador, listaCotas, listaCotasDesassociar);
 	}
 
 	private void processarEnderecos(Fiador fiador,
@@ -334,18 +355,23 @@ public class FiadorServiceImpl implements FiadorService {
 
 		for (EnderecoAssociacaoDTO enderecoAssociacao : listaEnderecoAssociacao) {
 
-			listaEndereco.add(enderecoAssociacao.getEndereco());
+			if (enderecoAssociacao.getEndereco() != null){
+				listaEndereco.add(enderecoAssociacao.getEndereco());
+			}
 
 			EnderecoFiador enderecoFiador = this.enderecoFiadorRepository.buscarEnderecoPorEnderecoFiador(enderecoAssociacao.getId(), fiador.getId());
 			
-			if (enderecoFiador != null){
+			if (enderecoFiador != null && enderecoFiador.getEndereco() != null){
 				idsEndereco.add(enderecoFiador.getEndereco().getId());
 				
 				this.enderecoFiadorRepository.remover(enderecoFiador);
 			}
 		}
 		
-		this.enderecoRepository.removerEnderecos(idsEndereco);
+		if (listaEndereco != null && !listaEndereco.isEmpty()){
+		
+			this.enderecoRepository.removerEnderecos(idsEndereco);
+		}
 	}
 	
 	private void processarTelefones(Fiador fiador, List<TelefoneAssociacaoDTO> listaTelefoneAdicionar,
@@ -565,6 +591,10 @@ public class FiadorServiceImpl implements FiadorService {
 	@Override
 	public Pessoa buscarPessoaFiadorPorId(Long idFiador) {
 		
+		if (idFiador == null){
+			throw new ValidacaoException(TipoMensagem.WARNING, "Id fiador é obrigatório.");
+		}
+		
 		Pessoa pessoa = this.fiadorRepository.buscarPessoaFiadorPorId(idFiador);
 		
 		if (pessoa instanceof PessoaFisica){
@@ -579,8 +609,19 @@ public class FiadorServiceImpl implements FiadorService {
 	public void excluirFiador(Long idFiador){
 		
 		Fiador fiador = this.fiadorRepository.buscarPorId(idFiador);
+
 		
 		if (fiador != null){
+			
+			if (fiador.getCotasAssociadas() != null){
+				
+				for (Cota cota : fiador.getCotasAssociadas()){
+					
+					cota.setFiador(null);
+					this.cotaRepository.alterar(cota);
+				}
+			}
+			
 			this.fiadorRepository.remover(fiador);
 		}
 	}
@@ -625,5 +666,42 @@ public class FiadorServiceImpl implements FiadorService {
 	public EnderecoFiador buscarEnderecoPorEnderecoFiador(Long idFiador, Long idEndereco) {
 		
 		return this.enderecoFiadorRepository.buscarEnderecoPorEnderecoFiador(idEndereco, idFiador);
+	}
+	
+	@Transactional(readOnly = true)
+	@Override
+	public Fiador obterFiadorPorId(Long idFiador){
+		
+		if (idFiador == null){
+			throw new ValidacaoException(TipoMensagem.WARNING, "Id fiador é obrigatório.");
+		}
+		
+		return this.fiadorRepository.buscarPorId(idFiador);
+	}
+
+	@Override
+	@Transactional(readOnly = true)
+	public List<Cota> obterCotasAssociadaFiador(Long idFiador) {
+		
+		if (idFiador == null){
+			throw new ValidacaoException(TipoMensagem.WARNING, "Id fiador é obrigatório.");
+		}
+		
+		return this.fiadorRepository.obterCotasAssociadaFiador(idFiador);
+	}
+
+	@Override
+	@Transactional(readOnly = true)
+	public boolean verificarAssociacaoFiadorCota(Long idFiador, Integer numeroCota) {
+		
+		if (idFiador == null){
+			throw new ValidacaoException(TipoMensagem.WARNING, "Id fiador é obrigatório.");
+		}
+		
+		if (numeroCota == null){
+			throw new ValidacaoException(TipoMensagem.WARNING, "Número cota é obrigatório.");
+		}
+		
+		return this.fiadorRepository.verificarAssociacaoFiadorCota(idFiador, numeroCota);
 	}
 }
